requirements.txt
# requirements.txt
Django>=4.0,<5.0
boto3>=1.20.0
django-ses>=3.0.0
python-decouple>=3.0